+++ src/Marker.php
@@ -18,17 +18,18 @@ use Symfony\UX\Map\Exception\InvalidArgumentException;
  *
  * @author Hugo Alliaume <hugo@alliau.me>
  */
-final readonly class Marker
+final readonly class Marker implements Element
 {
     /**
      * @param array<string, mixed> $extra Extra data, can be used by the developer to store additional information and
      *                                    use them later JavaScript side
      */
     public function __construct(
-        private Point $position,
-        private ?string $title = null,
-        private ?InfoWindow $infoWindow = null,
-        private array $extra = [],
+        public Point $position,
+        public ?string $title = null,
+        public ?InfoWindow $infoWindow = null,
+        public array $extra = [],
+        public ?string $id = null,
     ) {
     }
 
@@ -38,6 +39,7 @@ final readonly class Marker
      *     title: string|null,
      *     infoWindow: array<string, mixed>|null,
      *     extra: array,
+     *     id: string|null
      * }
      */
     public function toArray(): array
@@ -47,6 +49,7 @@ final readonly class Marker
             'title' => $this->title,
             'infoWindow' => $this->infoWindow?->toArray(),
             'extra' => $this->extra,
+            'id' => $this->id,
         ];
     }
 
@@ -56,6 +59,7 @@ final readonly class Marker
      *     title: string|null,
      *     infoWindow: array<string, mixed>|null,
      *     extra: array,
+     *     id: string|null
      * } $marker
      *
      * @internal