+++ CHANGELOG.md
@@ -8,6 +8,11 @@
    `HaversineDistanceCalculator`, `SphericalCosineDistanceCalculator` and `VincentyDistanceCalculator`.
 -  Add `CoordinateUtils` helper, to convert decimal coordinates (`43.2109`) in DMS (`56° 78' 90"`)
 
+-  Add property `$id` to `Marker`, `Polygon` and `Polyline` constructors
+-  Add method `Map::removeMarker(string|Marker $markerOrId)`
+-  Add method `Map::removePolygon(string|Polygon $polygonOrId)`
+-  Add method `Map::removePolyline(string|Polyline $polylineOrId)`
+
 ## 2.22
 
 -   Add method `Symfony\UX\Map\Renderer\AbstractRenderer::tapOptions()`, to allow Renderer to modify options before rendering a Map.